# gen/docs/swagger/swagger.yaml
definitions:
  domain.UserStatus:
    enum:
    - pending
    - active
    - locked
    - disabled
    type: string
    x-enum-varnames:
    - UserStatusPending
    - UserStatusActive
    - UserStatusLocked
    - UserStatusDisabled
  handlers.AuthLoginRequest:
    properties:
      device_id:
        type: string
      device_label:
        type: string
      identifier:
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
  handlers.AuthLoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      session:
        $ref: '#/definitions/handlers.SessionSummary'
      token_type:
        type: string
      user:
        $ref: '#/definitions/handlers.UserSummary'
    type: object
  handlers.AuthPendingResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/handlers.UserSummary'
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
      trace_id:
        type: string
    type: object
...
  handlers.JWKSKey:
    properties:
      alg:
        type: string
      e:
        type: string
      kid:
        type: string
      kty:
        type: string
      "n":
        type: string
      use:
        type: string
    type: object
  handlers.JWKSResponse:
    properties:
      keys:
        items:
          $ref: '#/definitions/handlers.JWKSKey'
        type: array
    type: object
...
  handlers.RegistrationRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handlers.RegistrationResponse:
    properties:
      delivery:
        type: string
      dev_code:
        description: Development only
        type: string
      dev_token:
        description: |-
          SECURITY: DevToken and DevCode are ONLY exposed in development mode
          In production, verification credentials are sent via secure channels
        type: string
      expires_at:
        type: string
      message:
        type: string
      requires_verification:
        type: boolean
      user:
        $ref: '#/definitions/handlers.UserSummary'
    type: object
...
  handlers.SessionSummary:
    properties:
      created_at:
        type: string
      device_label:
        type: string
      expires_at:
        type: string
      id:
        type: string
      last_seen:
        type: string
    type: object
...
  handlers.TokenRefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handlers.TokenRefreshResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/handlers.UserSummary'
    type: object
  handlers.UserSummary:
    properties:
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      roles:
        items:
          type: string
        type: array
      status:
        $ref: '#/definitions/domain.UserStatus'
      username:
        type: string
    type: object
  middleware.ProblemDetails:
    properties:
      detail:
        type: string
      instance:
        type: string
      retry_after:
        type: integer
      status:
        type: integer
      title:
        type: string
      trace_id:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
paths:
  /.well-known/jwks.json:
    get:
      description: Exposes the public keys used to verify IAM JWT signatures.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.JWKSResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Retrieve JSON Web Key Set
      tags:
      - Public
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Validates the provided identifier and password, returning access
        and refresh tokens on success.
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully authenticated
          schema:
            $ref: '#/definitions/handlers.AuthLoginResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Account pending verification or inactive
          schema:
            $ref: '#/definitions/handlers.AuthPendingResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/middleware.ProblemDetails'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service temporarily unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Authenticate a user with credentials
      tags:
      - Authentication
  /api/v1/auth/logout:
    post:
      description: Revokes the caller's active session using the access token's session
        context.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Logout the current session
      tags:
      - Authentication
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Issues a new access token and refresh token pair using a valid
        refresh token.
      parameters:
      - description: Refresh request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TokenRefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TokenRefreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Refresh an access token
      tags:
      - Authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user with the supplied credentials and contact information.
      parameters:
      - description: Registration request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegistrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.RegistrationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register a new user account
      tags:
      - Authentication
...
swagger: "2.0"